+++ TheEdge/core/src/com/filip/edge/screens/objects/EndTarget.java
@@ -9,12 +9,12 @@ import com.filip.edge.util.Constants;
 public class EndTarget extends AbstractCircleButtonObject {
     public static final String TAG = EndTarget.class.getName();
 
-    public EndTarget(int size, float x, float y, Color outsideColor, Color insideColor) {
+    public EndTarget(float size, float x, float y, Color outsideColor, Color insideColor) {
         super(size, x, y, outsideColor, insideColor);
     }
 
     @Override
-    public void fillInside(int size) {
-        buttonPixmap.fillCircle(size / 2, size / 2, (int) (size / 2 * Constants.END_CIRCLE_OUTLINE_RADIUS_MULTIPLIER));
+    public void fillInside(float size) {
+        buttonPixmap.fillCircle((int)(size / 2), (int)(size / 2), (int) (size / 2 * Constants.END_CIRCLE_OUTLINE_RADIUS_MULTIPLIER));
     }
 }
