+++ TheEdge/core/src/com/filip/edge/screens/objects/InfoButton.java
@@ -8,19 +8,19 @@ import com.badlogic.gdx.graphics.Color;
 public class InfoButton extends AbstractCircleButtonObject {
     public static final String TAG = InfoButton.class.getName();
 
-    public InfoButton(int size, float x, float y, Color outsideColor, Color insideColor) {
+    public InfoButton(float size, float x, float y, Color outsideColor, Color insideColor) {
         super(size, x, y, outsideColor, insideColor);
     }
 
     @Override
-    public void fillInside(int size) {
-        buttonPixmap.fillRectangle(size / 2 - size / 10,
-                size / 2 - (int) (size / 2.5),
-                size / 5,
-                size / 2);
-        buttonPixmap.fillRectangle(size / 2 - size / 10,
-                size / 2 + size / 5,
-                size / 5,
-                size / 5);
+    public void fillInside(float size) {
+        buttonPixmap.fillRectangle((int)(size / 2 - size / 10),
+                (int)(size / 2 - (size / 2.5)),
+                (int)(size / 5),
+                (int)(size / 2));
+        buttonPixmap.fillRectangle((int)(size / 2 - size / 10),
+                (int)(size / 2 + size / 5),
+                (int)(size / 5),
+                (int)(size / 5));
     }
 }
