+++ TheEdge/core/src/com/filip/edge/screens/objects/LeaderboardButton.java
@@ -8,28 +8,28 @@ import com.badlogic.gdx.graphics.Color;
 public class LeaderboardButton extends AbstractCircleButtonObject {
     public static final String TAG = LeaderboardButton.class.getName();
 
-    public LeaderboardButton(int size, float x, float y, Color outsideColor, Color insideColor) {
+    public LeaderboardButton(float size, float x, float y, Color outsideColor, Color insideColor) {
         super(size, x, y, outsideColor, insideColor);
     }
 
     @Override
-    public void fillInside(int size) {
+    public void fillInside(float size) {
         // middle
-        buttonPixmap.fillRectangle(size / 2 - size / 10,           // x
-                size / 2 - (int) (size / 5),    // y
-                size / 5,                       // width
-                size / 2);                      // height
+        buttonPixmap.fillRectangle((int)(size / 2 - size / 10),           // x
+                (int)(size / 2 - size / 5),    // y
+                (int)(size / 5),                       // width
+                (int)(size / 2));                      // height
         // right
-        buttonPixmap.fillRectangle(size / 2 + (int) (size / 6.5),  // x
-                size / 2 - (int) (size / 5),    // y
-                size / 5,                       // width
-                size / 4);                      // height
+        buttonPixmap.fillRectangle((int)(size / 2 + size / 6.5),  // x
+                (int)(size / 2 - size / 5),    // y
+                (int)(size / 5),                       // width
+                (int)(size / 4));                      // height
 
         // left
-        buttonPixmap.fillRectangle(size / 2 - size / 3,            // x
-                size / 2 - (int) (size / 5),    // y
-                size / 5,                       // width
-                size / 3);                      // height
+        buttonPixmap.fillRectangle((int)(size / 2 - size / 3),            // x
+                (int)(size / 2 - size / 5),    // y
+                (int)(size / 5),                       // width
+                (int)(size / 3));                      // height
 
 
     }
